# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RRHH API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/organizations": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "getOrganizationList",
                "description": "Get all Organizations",
                "operationId": "586ff2bb0a6c964dde6bf3ef4ec706c2",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/organizations": {
            "post": {
                "tags": [
                    "Organization"
                ],
                "summary": "createOrganization",
                "description": "Create Organization",
                "operationId": "00bdebd135b90132be55bb7a492e09ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/organizations/{id}": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "getOrganizationItem",
                "description": "Get Organization",
                "operationId": "70d53d114a77377c65553627a00177c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Organization",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Organization"
                ],
                "summary": "updateOrganization",
                "description": "Update Organization",
                "operationId": "cd18ac0a4a7093adff866a0c89241a5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Organization",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Organization"
                ],
                "summary": "deleteOrganization",
                "description": "Delete Organization",
                "operationId": "25d8ac9fd63f772ae17ed4917343a925",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Organization",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    }
}